requirements/requirements.txt
matplotlib==3.5.2
scikit-learn>=1.1.1
scipy>=1.8.1
omegaconf>=2.2.2
pandas>=1.4.2
tqdm>=4.64.0
numpy>=1.22.0
sqlalchemy>=1.4.37
vertica-python==1.1.0
sqlalchemy-vertica-python>=0.5.10
networkx>=2.8.4
hydra-core>=1.2.0
pytorch-lightning>=1.6.4
torchmetrics>=0.9.1
torch>=1.11.0
tqdm>=4.64.0
tensorboard==2.11.2
pyarrow==11.0.0
loguru==0.6.0
